+++ CHANGES.rst
@@ -67,6 +67,7 @@ Internal Changes
 - Disallow server-side requests to private, loopback, reserved and link-local IP ranges in
   places where the URL is user-provided (Mastodon URL check, LaTeX image retrieval, static
   site generation) (:pr:`7244`)
+- Log requests to the legacy export API to ``indico.log`` (:pr:`7290`)
 
 
 Version 3.3.9